Add files via upload
Copy of Masters thesis, submitted as part of the MSc in Data Science, Machine Learning & Advanced Computing. The Executive Summary is as follows:

The aim of this project is to develop tools that can offer recommendations to firms for technology areas that they might consider for new R&D. The recommendations should be based on the firms existing patent portfolio, and will exploit the massive patent database of the US Patent and Trademark Office (USPTO) to reveal the most likely productive for the firm to move into, based on the behaviour of other inventors/firms. The literature argues that higher value patents come from recombinative innovation, in which different technologies are combined in new ways. It is also expected that this is true only to a certain degree of radically (the conceptual distance between the technologies being mixed), beyond which the innovators themselves, or the market cannot absorb the ideas fully and the innovation has lower value. These two forces together create an expected inverted U-shape curve. This project thus aimed to: firstly identify the optimal technology distance to get the greatest expected patent value; and secondly, use this distance to recommend specific technologies to a firm. The project originally used the NBER dataset for 1997-2006. These data have been
used for many studies, and included data cleaning of patent assignees. The results are applied to an example, showing how they can be used to suggest future technological directions for a computer company.