Add files via upload
A python Jupyter Notebook, containing an initial experiment to auto-code the medical texts based on the dataset of human-coded texts generated earlier. On a set of test labels, we obtain an accuracy of over 90%, using an SVM classifier.
This notebook has been heavily edited and has certain cells deactivated to ensure that no personal data from the medical texts is revealed.